refact : 특정일자 경기 가져오기 분리 #72
