+++ src/main/java/com/inninglog/inninglog/domain/home/usecase/HomeUsecase.java
@@ -6,7 +6,7 @@ import com.inninglog.inninglog.domain.kbo.domain.Game;
 import com.inninglog.inninglog.domain.kbo.dto.gameReport.GameHomeResDto;
 import com.inninglog.inninglog.domain.kbo.dto.gameReport.WinningRateResult;
 import com.inninglog.inninglog.domain.kbo.service.GameReportService;
-import com.inninglog.inninglog.domain.kbo.service.GameValidateService;
+import com.inninglog.inninglog.domain.kbo.service.GameGetService;
 import com.inninglog.inninglog.domain.member.domain.Member;
 import com.inninglog.inninglog.domain.member.service.MemberValidateService;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +21,7 @@ public class HomeUsecase {
 
     private final MemberValidateService memberValidateService;
     private final GameReportService gameReportService;
-    private final GameValidateService gameValidateService;
+    private final GameGetService gameGetService;
     private final HomeService homeService;
 
     @Transactional(readOnly = true)
@@ -29,7 +29,7 @@ public class HomeUsecase {
         Member member = memberValidateService.findById(memberId);
         memberValidateService.validateTeam(member);
         WinningRateResult winningRateResult = gameReportService.forHomeCaculateWin(member);
-        List<Game> games = gameValidateService.findByTeam(member.getTeam().getId());
+        List<Game> games = gameGetService.findByTeam(member.getTeam().getId());
         List<GameHomeResDto> myTeamSchedule = homeService.getAllGamesForTeam(games, member.getTeam().getId());
 
         return HomeResDTO.from(member,winningRateResult.getWinningRateHalPoongRi(), myTeamSchedule);
