+++ src/main/java/com/inninglog/inninglog/domain/kbo/service/GameGetService.java
@@ -9,12 +9,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class GameValidateService {
+public class GameGetService {
 
     private final GameRepository gameRepository;
 
@@ -38,5 +39,16 @@ public class GameValidateService {
                     log.warn("⚠️ 존재하지 않는 경기: gameId={}", gameId);
                     return new CustomException(ErrorCode.GAME_NOT_FOUND);
                 });
+    }
+
+    @Transactional(readOnly = true)
+    public Game findByDateAndTeamId(LocalDate gameDate, Long teamId){
+        Game game = gameRepository.findByDateAndTeamId(gameDate, teamId);
+        if (game == null) {
+            log.warn("⚠️ 해당 날짜에 경기 없음: date={}, teamId={}", gameDate, teamId);
+            return null;
+        }
+
+        return game;
     }
 }
